add event component and total amount added
