+++ client/src/pages/Profile/Profile.jsx
@@ -9,10 +9,12 @@ const { Step } = Steps;
 const Profile = () => {
   const [currentCompletion, setCurrent] = useState(0)
   const [err, setErr] = useState(0)
+  const [totalAmt,settotalAmt ] = useState(0)
   useEffect(() => {
     axios.get('http://localhost:5000/api/users/profile-details', {
       headers: { Authorization: `bearer ${localStorage.getItem('usertoken')}` }
     }).then((response) => {
+      settotalAmt(response.data.moneyOwed)
       let isCriteria = Object.values(response.data.criteria).every(val => val === true)
       if (response.data.hasFilledProfile) {
         setCurrent(1)
@@ -26,6 +28,7 @@ const Profile = () => {
     })
       .catch(e => setErr(1));
   }, []);
+  const paidDescription=`Pay ${totalAmt} ₹ as total entry fee of participation in events`
   return (
 
     <Fragment>
@@ -43,7 +46,7 @@ const Profile = () => {
                 <Steps current={currentCompletion} responsive={true} direction={"vertical"}>
                   <Step title="Update Profile" description="Fill the form to update your name and phone no." />
                   <Step title="Meet Criterion" description="Register for events! Criteria is not yet satisfied" />
-                  <Step title="Paid" description="Pay whole amount you got on your profile" />
+                  <Step title="Paid" description={paidDescription} />
                 </Steps>
               </div>
             </div>
